feat: create ColumnConfigurations to encapsulate functionality
